// src/workspace/settings/__generated__/MaxJobDurationSettingUpdateMutation.graphql.ts
export type ProblemType = "BAD_REQUEST" | "CONFLICT" | "FORBIDDEN" | "NOT_FOUND" | "%future added value";
export type UpdateWorkspaceInput = {
  clientMutationId?: string | null;
  description?: string | null;
  maxJobDuration?: number | null;
  metadata?: ResourceMetadataInput | null;
  terraformVersion?: string | null;
  workspacePath: string;
};
export type ResourceMetadataInput = {
  version: string;
};
export type MaxJobDurationSettingUpdateMutation$variables = {
  input: UpdateWorkspaceInput;
};
export type MaxJobDurationSettingUpdateMutation$data = {
  readonly updateWorkspace: {
    readonly problems: ReadonlyArray<{
      readonly field: ReadonlyArray<string> | null;
      readonly message: string;
      readonly type: ProblemType;
    }>;
    readonly workspace: {
      readonly maxJobDuration: number;
    } | null;
  };
};
export type MaxJobDurationSettingUpdateMutation = {
  response: MaxJobDurationSettingUpdateMutation$data;
  variables: MaxJobDurationSettingUpdateMutation$variables;
};